Support Windows partially
Only supports the `-o, --output DIR` option and the
`installed td-agent-gem list` collection for Fluentd (fluent-package,
td-agent).

Splitting the implementation just to easily provide minimal support for
Windows. If we are going to support more features for Windows, it would
be better to remove "diagutils-win.rb" and improve the origin file
"diagutils.rb".

Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>